Add specific cell subclasses to simplify instantiation
