+++ ALES Project/src/creature/cells/CoreCell.java
@@ -1,0 +1,19 @@
+/*
+ * To change this license header, choose License Headers in Project Properties.
+ * To change this template file, choose Tools | Templates
+ * and open the template in the editor.
+ */
+
+package creature.cells;
+
+/**
+ *
+ * @author bhargav
+ */
+    public class CoreCell extends Cell{
+
+        public CoreCell(int x, int y) {
+            
+            super(1, 6, 4, x, y);
+        }
+    }
