+++ ALES Project/src/creature/cells/DeadCell.java
@@ -1,0 +1,19 @@
+/*
+ * To change this license header, choose License Headers in Project Properties.
+ * To change this template file, choose Tools | Templates
+ * and open the template in the editor.
+ */
+
+package creature.cells;
+
+/**
+ *
+ * @author bhargav
+ */
+public class DeadCell extends Cell {
+
+        public DeadCell(int x, int y) {
+
+            super(6, 0, 0, x, y);
+        }
+    }
