+++ ALES Project/src/creature/cells/HunterCell.java
@@ -1,0 +1,26 @@
+/*
+ * To change this license header, choose License Headers in Project Properties.
+ * To change this template file, choose Tools | Templates
+ * and open the template in the editor.
+ */
+
+package creature.cells;
+
+/**
+ *
+ * @author bhargav
+ */
+    public class HunterCell extends Cell{
+
+        public HunterCell(int x, int y) {
+            
+            super(3, 2, 1, x, y);
+        }
+
+//        @Override
+//        public void doAction() {
+//
+//            //creature.use energy -USER VAR-
+//            //ask map to return int of energy gained (creature asks map to do damage)
+//        }
+    }
