+++ ALES Project/test/mapTest/TerrainTest.java
@@ -6,8 +6,11 @@
 
 package mapTest;
 
-import creature.Cell;
+import creature.cells.Cell;
 import creature.Creature;
+import creature.cells.StructureCell;
+import java.util.ArrayList;
+import java.util.List;
 import map.Terrain;
 
 /**
@@ -18,9 +21,17 @@ public class TerrainTest {
     public static void main(String[] args) {
         int[][] env = new int[20][20];
         Cell[][] cellArray = new Cell[Creature.SIDE_LENGTH][Creature.SIDE_LENGTH];
-        Terrain map = new Terrain(env);
+        for (int row = 0; row<cellArray.length; row++){
+            for(int c = 0; c<cellArray[0].length; c++){
+                cellArray[row][c] = new StructureCell(row,c);
+            }
+        }
+        List<Creature> pop = new ArrayList<Creature>();
         Creature test1 = new Creature(cellArray, 0, null, 3, 3);
         Creature test2 = new Creature(cellArray, 0, null, 2, 2);
+        pop.add(test1);
+        pop.add(test2);
+        Terrain map = new Terrain(env,pop);
         Cell c = map.cellAtAbsPos(4, 4);
         System.out.println(c);
     }
